Move the table creation closer to front
