+++ containers/ecr-viewer/src/app/view-data/components/AccordionContainer.tsx
@@ -84,7 +84,7 @@ const AccordianContainer: React.FC<AccordionContainerProps> = ({
               clinicalData.immunizationsDetails.availableData
             }
             treatmentData={clinicalData.treatmentData.availableData}
-            planOfTreatment={clinicalData.planOfTreatment.availableData}
+            planOfTreatment={clinicalData.planOfTreatment}
           />
         </>
       ),
